Fixes issue #4 association inconsistency.
`rel.formal_attributes` now determines the order in which matches are found.
It seems as if `linkedGroups` order is not always the same, whereas the
order of `rel.formal_attributes` order is always the same. So reversing
the loops makes that the order in which matches are found is now determined
by the order of `rel.formal_attributes` as opposed to `linkedGroups`.

Also in `match` there was an issue in which it was possible that `ret` was
not changed from its `None` initialization, causing `append` to be called on
a `None` object resulting in an exception. Also turned the order of the tests
around, because sometimes an empty list (or `None` before the `ret` fix was
returned when `adr` was in `mdict` and would have returned a value if this
test was first, which it now is.